Specify working version of Rust
The Rust target wasm32-wasi has been renamed to wasm32-wasip1 in recent versions of Rust. Using the original setup results in this error: 0.821 error: toolchain 'stable-x86_64-unknown-linux-gnu' does not support target 'wasm32-wasi'; did you mean 'wasm32-wasip1'?

This is an attempt to install a version of rust that existed when this project was last functional.